pyproject.toml
[tool.poetry]
name = "snaptext-ai"
version = "0.1.0"
description = ""
authors = ["Rupanshu Kapoor <rupanshukapoor@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.2.7"
streamlit = "^1.36.0"
streamlit-chat = "^0.1.1"
google-cloud-documentai = "^2.29.2"
pymupdf = "^1.24.7"
firebase-admin = "^6.5.0"
pyrebase4 = "^4.8.0"
google-cloud-firestore = "^2.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
